+++ CHANGELOG.md
@@ -6,6 +6,7 @@ ChangeLog
 ------------------
 
 - IMPROVEMENT: rewrite "wav" node to be self-contained and preserve chunk information (time, meta)
+- IMPROVEMENT: the "trace" node now also outputs meta information
 - UPGRADE: upgrade NPM dependencies
 
 1.1.0 (2025-07-19)