+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ChangeLog
 - IMPROVEMENT: add "gender" node for male/female speaker detection (result is passed as meta information)
 - IMPROVEMENT: add "filter" node for filtering based on meta information
 - IMPROVEMENT: do not segment audio stream chunks for "vad" node even if it internally has to segment it
+- IMPROVEMENT: allow chunking duration/size to be controlled for device/file nodes
 - UPGRADE: upgrade NPM dependencies
 
 1.1.0 (2025-07-19)